Snapshot from ross-pro-transfer-20260908 for GitHub mirror.
Includes local training/eval scripts and model code updates; excludes data/, checkpoints/, and outputs/.

+++ MMVP/mmvp_eval.py
@@ -87,7 +87,9 @@ def eval_model(args):
     # disable_torch_init()  # DO NOT ENABLE THIS: KILLS PERFORMANCE
     model_path = os.path.expanduser(args.model_path)
     model_name = get_model_name_from_path(model_path)
-    tokenizer, model, image_processor, context_len = load_pretrained_model(model_path, args.model_base, model_name)
+    tokenizer, model, image_processor, context_len = load_pretrained_model(
+        model_path, args.model_base, model_name, torch_dtype=torch.float16,
+    )
 
     images = {}
     for i in range(300):
